allow -1 to indicate voltage not yet stable

+++ src/main/sensors/battery.h
@@ -78,7 +78,7 @@ uint16_t getBatteryCurrentSample(void);
 uint16_t getLegacyBatteryCurrent(void);
 uint16_t getBatteryCapacity(void);
 uint32_t getBatteryCapacityUsed(void);
-uint8_t getBatterySmartFuel(void);
+int getBatterySmartFuel(void);
 
 uint8_t calculateBatteryPercentageRemaining(void);
